# Conflicts:
#	Cargo.lock
#	Cargo.toml
#	netwerk/protocol/http/Http3Session.cpp
#	netwerk/protocol/http/nsHttpChannel.cpp
#	netwerk/protocol/webtransport/WebTransportSessionProxy.cpp
#	netwerk/test/http3server/Cargo.toml
#	netwerk/test/http3server/src/main.rs
#	netwerk/test/unit/xpcshell.toml
#	third_party/rust/neqo-http3/.cargo-checksum.json
#	third_party/rust/neqo-http3/Cargo.toml
#	third_party/rust/neqo-transport/.cargo-checksum.json
#	third_party/rust/neqo-transport/Cargo.toml
#	third_party/rust/test-fixture/.cargo-checksum.json
#	third_party/rust/test-fixture/Cargo.toml
#	third_party/rust/test-fixture/src/lib.rs

+++ image/decoders/nsJXLDecoder.cpp
@@ -3,15 +3,14 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-#include "ImageLogging.h"  // Must appear first
-
 #include "nsJXLDecoder.h"
 
 #include "AnimationParams.h"
-#include "mozilla/CheckedInt.h"
-#include "gfxPlatform.h"
+#include "ImageLogging.h"  // Must appear first
 #include "RasterImage.h"
 #include "SurfacePipeFactory.h"
+#include "gfxPlatform.h"
+#include "mozilla/CheckedInt.h"
 
 using namespace mozilla::gfx;
 
@@ -418,7 +417,7 @@ nsresult nsJXLDecoder::AllocateFrameBuffers() {
   }
 
   if (mPixelFormat.value() == PixelFormat::Cmyk8 &&
-      !mKBuffer.resize(size.width * size.height)) {
+      !mKBuffer.resize(size_t(size.width) * size.height)) {
     return NS_ERROR_FAILURE;
   }
